add naga validator, build docs for various targets
